JPIE : Corrections de bugs : - Pb de modification de la réservation d'un produit d'abonnement par un adhérent => La select box doit maintenant être inactive - Impossible de modifier la date de naissance si il n'y a pas de date d'adhésion - Dans les liste Tri sur le N° adhérent/Ferme/Producteur ne fonctionne pas - Edition du marché : format des chiffres de quantité des réservation non conforme - Maintenance non suppression de la configuration mail et de la configuration .htaccess

+++ Maintenance/Maj/deploiementFile.php
@@ -5,7 +5,8 @@
 function supprimerDossier($pPath) {
 	$d = dir($pPath);
 	while (false !== ($entry = $d->read())) {	   
-	   if(	$entry != '.' && $entry != '..' && $entry != 'index.php' && $entry != 'Maintenance.php' && $entry != 'Maintenance' && $entry != 'logs' && $entry != "DB.php") {
+	   if(	$entry != '.' && $entry != '..' && $entry != 'index.php' && $entry != 'Maintenance.php' && $entry != 'Maintenance' && $entry != 'logs' && $entry != "DB.php" && $entry != "Mail.php"
+	    && $entry != ".htaccess") {
    		if(is_dir($d->path.'/'.$entry)) {
    			supprimerDossier($d->path.'/'.$entry);
 			if($entry != 'configuration') {
@@ -25,7 +26,8 @@ supprimerDossier(DOSSIER_SITE);
 function parcourirDossier($pPathIn,$pPathOut) {
 	$d = dir($pPathIn);
 	while (false !== ($entry = $d->read())) {	   
-	   if(	$entry != '.' && $entry != '..' && $entry != 'index.php' && $entry != 'Maintenance.php' && $entry != 'Maintenance' && $entry != 'update.sql' && $entry != "DB.php") {
+	   if(	$entry != '.' && $entry != '..' && $entry != 'index.php' && $entry != 'Maintenance.php' && $entry != 'Maintenance' && $entry != 'update.sql' && $entry != "DB.php" && $entry != "Mail.php"
+	   && $entry != ".htaccess") {
    		if(is_dir($d->path.'/'.$entry)) {
 			if(!is_dir($pPathOut .'/'. $entry)) {
 				mkdir($pPathOut .'/'. $entry);
